Fix monitor sending duplicated data when client connects
The server is supposed to send a limited number of lines from the log
once the client is connected; however a bug made the server double
such lines for each new connection, and quickly reach a large number of
duplicates, putting the client on its knees (to give a concrete
example, after just 11 conections the client will receive 10.240 lines
lines istead of 10, that is basically the whole content of a log file
of ~900K. With 20 connections the lines received are 5.242.880).

- Add env variable for WebSocket's host address

If REACT_APP_HOST_WS is set during hornet's build, it will supersed
the current host of the webpage for the WebSocket. It's useful during
development of the front-end, to connect to a remote machine for data
fetching.

- Fix Net Latency color palette
- ESLint / Prettier standard code formatting
- Fix scrolling on charts